requirements.txt
pydantic-ai
openai
python-dotenv
logfire
requests
qdrant-client
fastapi
uvicorn[standard]
